Remove duplicate Chaos discovery source
